+++ javaparser-core-testing/src/test/java/com/github/javaparser/ast/stmt/ForEachStmtTest.java
@@ -4,7 +4,7 @@ import com.github.javaparser.ast.type.ClassOrInterfaceType;
 import com.github.javaparser.ast.type.PrimitiveType;
 import org.junit.jupiter.api.Test;
 
-import static com.github.javaparser.JavaParser.parseStatement;
+import static com.github.javaparser.StaticJavaParser.parseStatement;
 import static org.junit.jupiter.api.Assertions.*;
 
 class ForEachStmtTest {