+++ javaparser-core-testing/src/test/java/com/github/javaparser/ast/visitor/ModifierVisitorTest.java
@@ -21,7 +21,6 @@
 
 package com.github.javaparser.ast.visitor;
 
-import com.github.javaparser.JavaParser;
 import com.github.javaparser.ast.NodeList;
 import com.github.javaparser.ast.body.BodyDeclaration;
 import com.github.javaparser.ast.body.VariableDeclarator;
@@ -30,7 +29,8 @@ import com.github.javaparser.ast.expr.IntegerLiteralExpr;
 import com.github.javaparser.ast.expr.StringLiteralExpr;
 import org.junit.jupiter.api.Test;
 
-import static com.github.javaparser.JavaParser.parseExpression;
+import static com.github.javaparser.StaticJavaParser.parseBodyDeclaration;
+import static com.github.javaparser.StaticJavaParser.parseExpression;
 import static com.github.javaparser.utils.Utils.EOL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -101,7 +101,7 @@ class ModifierVisitorTest {
 
     @Test
     void fieldDeclarationCantSurviveWithoutVariables() {
-        final BodyDeclaration<?> bodyDeclaration = JavaParser.parseBodyDeclaration("int x=1;");
+        final BodyDeclaration<?> bodyDeclaration = parseBodyDeclaration("int x=1;");
 
         final Visitable result = bodyDeclaration.accept(new ModifierVisitor<Void>() {
             public Visitable visit(VariableDeclarator x, Void arg) {
@@ -114,7 +114,7 @@ class ModifierVisitorTest {
 
     @Test
     void variableDeclarationCantSurviveWithoutVariables() {
-        final BodyDeclaration<?> bodyDeclaration = JavaParser.parseBodyDeclaration("void x() {int x=1;}");
+        final BodyDeclaration<?> bodyDeclaration = parseBodyDeclaration("void x() {int x=1;}");
 
         final Visitable result = bodyDeclaration.accept(new ModifierVisitor<Void>() {
             public Visitable visit(VariableDeclarator x, Void arg) {