+++ javaparser-core-testing/src/test/java/com/github/javaparser/printer/lexicalpreservation/transformations/ast/body/StatementTransformationsTest.java
@@ -25,12 +25,13 @@ import java.io.IOException;
 
 import org.junit.jupiter.api.Test;
 
-import com.github.javaparser.JavaParser;
 import com.github.javaparser.ast.NodeList;
 import com.github.javaparser.ast.expr.NameExpr;
 import com.github.javaparser.ast.stmt.Statement;
 import com.github.javaparser.printer.lexicalpreservation.AbstractLexicalPreservingTest;
 import com.github.javaparser.printer.lexicalpreservation.LexicalPreservingPrinter;
+
+import static com.github.javaparser.StaticJavaParser.parseStatement;
 
 /**
  * Transforming Statement and verifying the LexicalPreservation works as expected.
@@ -38,7 +39,7 @@ import com.github.javaparser.printer.lexicalpreservation.LexicalPreservingPrinte
 class StatementTransformationsTest extends AbstractLexicalPreservingTest {
 
     Statement consider(String code) {
-        Statement statement = JavaParser.parseStatement(code);
+        Statement statement = parseStatement(code);
         LexicalPreservingPrinter.setup(statement);
         return statement;
     }