+++ javaparser-core/src/main/java/com/github/javaparser/ast/expr/Name.java
@@ -20,7 +20,6 @@
  */
 package com.github.javaparser.ast.expr;
 
-import com.github.javaparser.JavaParser;
 import com.github.javaparser.TokenRange;
 import com.github.javaparser.ast.AllFieldsConstructor;
 import com.github.javaparser.ast.Node;
@@ -33,10 +32,9 @@ import com.github.javaparser.metamodel.NameMetaModel;
 import com.github.javaparser.metamodel.NonEmptyProperty;
 import com.github.javaparser.metamodel.OptionalProperty;
 import java.util.Optional;
+
 import static com.github.javaparser.utils.Utils.assertNonEmpty;
 import com.github.javaparser.ast.visitor.CloneVisitor;
-import com.github.javaparser.metamodel.NameMetaModel;
-import com.github.javaparser.metamodel.JavaParserMetaModel;
 import com.github.javaparser.ast.Generated;
 
 /**
@@ -109,20 +107,6 @@ public final class Name extends Node implements NodeWithIdentifier<Name> {
         notifyPropertyChange(ObservableProperty.IDENTIFIER, this.identifier, identifier);
         this.identifier = identifier;
         return this;
-    }
-
-    /**
-     * Creates a new {@link Name} from a qualified name.<br>
-     * The qualified name can contains "." (dot) characters.
-     *
-     * @param qualifiedName qualified name
-     * @return instanceof {@link Name}
-     * @deprecated use JavaParser.parseName instead
-     */
-    @Deprecated
-    public static Name parse(String qualifiedName) {
-        assertNonEmpty(qualifiedName);
-        return JavaParser.parseName(qualifiedName);
     }
 
     /**