+++ javaparser-symbol-solver-testing/src/test/java/com/github/javaparser/symbolsolver/Issue347.java
@@ -1,6 +1,5 @@
 package com.github.javaparser.symbolsolver;
 
-import com.github.javaparser.JavaParser;
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.body.FieldDeclaration;
 import com.github.javaparser.resolution.types.ResolvedType;
@@ -12,7 +11,9 @@ import com.github.javaparser.symbolsolver.resolution.typesolvers.ReflectionTypeS
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import static com.github.javaparser.StaticJavaParser.parse;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class Issue347 extends AbstractResolutionTest{
 
@@ -34,11 +35,11 @@ class Issue347 extends AbstractResolutionTest{
                 "class UsingFoo {\n" +
                 "    Foo myFooField;\n" +
                 "}";
-        CompilationUnit cu = JavaParser.parse(code);
+        CompilationUnit cu = parse(code);
         FieldDeclaration fieldDeclaration = Navigator.findNodeOfGivenClass(cu, FieldDeclaration.class);
         ResolvedType fieldType = javaParserFacade.getType(fieldDeclaration);
-        assertEquals(true, fieldType.isReferenceType());
-        assertEquals(true, fieldType.asReferenceType().getTypeDeclaration().isEnum());
+        assertTrue(fieldType.isReferenceType());
+        assertTrue(fieldType.asReferenceType().getTypeDeclaration().isEnum());
         assertEquals("foo.bar.Foo", fieldType.asReferenceType().getQualifiedName());
     }
 }